+++ pkg/ast/pipesearch/searchHandler_test.go
@@ -47,51 +47,3 @@ func Test_parseSearchBody(t *testing.T) {
 	assert.Equal(t, uint64(700), finalSize, "expected=%v, actual=%v", 700, scroll)
 	assert.Equal(t, 500, scroll, "expected=%v, actual=%v", 500, scroll)
 }
-
-func Test_parseAlphaNumTime(t *testing.T) {
-
-	nowTs := uint64(1659874108987)
-
-	defValue := uint64(12345)
-
-	inp := "now"
-	expected := nowTs
-	actual := parseAlphaNumTime(nowTs, inp, defValue)
-	assert.Equal(t, expected, actual, "expected=%v, actual=%v", expected, actual)
-
-	inp = "now-1m"
-	expected = nowTs - 1*60_000
-	actual = parseAlphaNumTime(nowTs, inp, defValue)
-	assert.Equal(t, expected, actual, "expected=%v, actual=%v", expected, actual)
-
-	inp = "now-12345m"
-	expected = nowTs - 12345*60_000
-	actual = parseAlphaNumTime(nowTs, inp, defValue)
-	assert.Equal(t, expected, actual, "expected=%v, actual=%v", expected, actual)
-
-	inp = "now-1h"
-	expected = nowTs - 1*3600_000
-	actual = parseAlphaNumTime(nowTs, inp, defValue)
-	assert.Equal(t, expected, actual, "expected=%v, actual=%v", expected, actual)
-
-	inp = "now-365d"
-	expected = nowTs - 365*24*3600*1_000
-	actual = parseAlphaNumTime(nowTs, inp, defValue)
-	assert.Equal(t, expected, actual, "expected=%v, actual=%v", expected, actual)
-
-	inp = "invalidTimeString"
-	expected = 12345
-	actual = parseAlphaNumTime(nowTs, inp, defValue)
-	assert.Equal(t, expected, actual, "expected=%v, actual=%v", expected, actual)
-
-	inp = "now-365x"
-	expected = 12345
-	actual = parseAlphaNumTime(nowTs, inp, defValue)
-	assert.Equal(t, expected, actual, "expected=%v, actual=%v", expected, actual)
-
-	inp = "now-5xm"
-	expected = 12345
-	actual = parseAlphaNumTime(nowTs, inp, defValue)
-	assert.Equal(t, expected, actual, "expected=%v, actual=%v", expected, actual)
-
-}
